+++ typist/src/main/java/com/karuslabs/typist/generation/chunks/Header.java
@@ -61,7 +61,7 @@ public record Header() {
                import org.bukkit.command.CommandSender;
 
                /**
-                * This file was generated at %s using Chimera %s
+                * This file was generated at %s using Typist %s
                 */
                """.formatted(LocalDateTime.now(), VERSION));
     }